Kernel: Introduce SpinlockProtected
